// mobile/Friendlinq_frontend/src/app/components/models/signup.model.ts
export class SignUp {
  fullName: string;
  email: string;
  password: string;
  age: number;
  country: string;
  gender: string;
  deviceType: string;
  deviceToken: string;
}
export class Profile {
  fullName: string;
  country: string;
  gender: string;
  city: string;
  state: string;
  phoneNumber: string;
  about: string;
}